polish(eval): apply review findings on Issue #52 PR1
Phase 1 (parallel review: simplify + Codex + CodeRabbit + Gemini):
- Codex / CodeRabbit: 0 findings
- Gemini: capacity exhausted on gemini-2.5-pro, skipped per polish protocol
- simplify (3 sub-reviewers): 12 findings, 6 applied, 6 deferred

Applied:
- REUSE-001 / EFF-001: drop redundant `corpus_ids` HashSet; use
  `corpus_index.contains_key` instead. Hoists corpus_index construction
  before the validation loop so the same lookup serves both stages.
- CQ-5: extract `validate_relevance_map_against_corpus` helper.
  evaluate_oracle was 58 lines, now ~30. Validation rationale (silent
  drop in apply_reranker → vacuous AC 4 gate) lives on the helper.
- EFF-003: replace per-call `HashSet<&str>` with `.iter().any()`. Typical
  relevant-doc count is 1-5; linear scan is cheaper than HashSet alloc.
- CQ-4: drop stale `#[allow(clippy::too_many_arguments)]` on 5-arg
  `run_capture_baseline_with` / `run_capture_oracle_with`. Clippy default
  threshold is 7.
- CQ-6: split tests to `src/eval/oracle_pipeline/tests.rs` per
  `rules/development/STRUCTURE.md` (oracle_pipeline.rs was 642 lines
  > 400 ceiling; tests dominated). Production code now ~245 lines.

Phase 2 (enhancer-code): trimmed slop comments — scope paragraph
restating headers, AC-restating test docblocks, redundant Issue #52
parentheticals on internal helpers. Preserved every WHY comment
(f64::MAX precision quirk, vacuous AC 4 risk, project T-NNN convention).

Deferred (PR2 candidates):
- CQ-1: replace `_oracle` string suffix with `EvaluateKind` enum + bool
- CQ-2 / REUSE-002: unify `dispatch_pipeline` helpers
- CQ-3 / REUSE-003: unify capture handlers
- EFF-002: hoist `merge_config.clone()` per query

Gates: cargo test --all-features (165 + 7 + 18 doctests pass), cargo
clippy --all-targets --all-features -- -D warnings (clean), cargo fmt
-- --check (clean).
